+++ src/Catty/SensorManager.m
@@ -40,8 +40,6 @@ NSString * const sensorStringArray[] = {
     @"OBJECT_LAYER",
 };
 
-
-
 +(Sensor) sensorForString:(NSString*)sensor
 {
     if([sensor isEqualToString:@"X_ACCELERATION"]) {
@@ -92,18 +90,17 @@ NSString * const sensorStringArray[] = {
     return -1;
 }
 
-+ (NSString*) stringForSensor:(Sensor)sensor
++ (NSString*)stringForSensor:(Sensor)sensor
 {
-    if(sensor < sizeof(sensorStringArray) / sizeof(Sensor))
+    if (((NSInteger) sensor) < ((NSInteger)(sizeof(sensorStringArray) / sizeof(Sensor))))
         return sensorStringArray[sensor];
-    
+
     return @"Unknown Sensor";
 }
 
-+(BOOL) isObjectSensor:(Sensor)sensor
++ (BOOL)isObjectSensor:(Sensor)sensor
 {
     return (sensor >= OBJECT_X && sensor <= OBJECT_LAYER) ? YES : NO;
 }
-
 
 @end
